Don't allow only one key at a time
This does work but it's a bit jank
Honestly I don't think this'll cause any issues unless you are trying to break it

e.g. The sequence:
"Press 1", "Press 2", "Release 2", "Release 1"
Will type "121", with the second erroneous 1 when you "Release 2"
But the OS doesn't require any button holds, is this ever going to come up?